helloworld/src/main/java/com/helloworld/repository/MemoryMemberRespository.java
package com.helloworld.repository;

import com.helloworld.domain.Member;

import java.util.*;

public class MemoryMemberRespository implements MemberRespository {

    private static Map<Long, Member> store = new HashMap<>();
    private static long sequence = 0L;


    @Override
    public Member save(Member member) {
        member.setId(++sequence);
        store.put(member.getId(),member);
        return member;
    }

    @Override
    public Optional<Member> findById(Long id) {
        return Optional.ofNullable(store.get(id));   // 값이 null 일때는 Optional.nullable 로 감쌈
    }

    @Override
    public Optional<Member> findByName(String name) {
        return store.values().stream()
                .filter(member -> member.getName().equals(name))
                .findAny();
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    // store를 비움
    public void clearStore() {
        store.clear();
    }
}
